New design for diagnostic tab
