+++ locales/en/plugin__kubevirt-plugin.json
@@ -405,6 +405,7 @@
   "Enable Descheduler": "Enable Descheduler",
   "Enable headless mode": "Enable headless mode",
   "Enable preallocation": "Enable preallocation",
+  "Enabled": "Enabled",
   "Ensure your PVC size covers the requirements of the uncompressed image and any other space requirements.": "Ensure your PVC size covers the requirements of the uncompressed image and any other space requirements.",
   "Enter a container image (for example: {{containerDisk}})": "Enter a container image (for example: {{containerDisk}})",
   "Enter key=value": "Enter key=value",
@@ -517,7 +518,6 @@
   "Last 5 minutes": "Last 5 minutes",
   "Last 6 hours": "Last 6 hours",
   "Last restored": "Last restored",
-  "Last transition time": "Last transition time",
   "Launch Migration Toolkit for Virtualization web console": "Launch Migration Toolkit for Virtualization web console",
   "Launch Remote Desktop": "Launch Remote Desktop",
   "Launch Remote Viewer": "Launch Remote Viewer",
@@ -907,6 +907,7 @@
   "Start this VirtualMachine in pause mode": "Start this VirtualMachine in pause mode",
   "Start VirtualMachine on clone": "Start VirtualMachine on clone",
   "Status": "Status",
+  "Status Conditions": "Status Conditions",
   "Stop": "Stop",
   "Stop VirtualMachine": "Stop VirtualMachine",
   "Storage": "Storage",
@@ -1117,6 +1118,7 @@
   "Volume Mode": "Volume Mode",
   "Volume name": "Volume name",
   "Volume parameters": "Volume parameters",
+  "Volume Snapshot Status": "Volume Snapshot Status",
   "Volumes project": "Volumes project",
   "Volumes to boot from": "Volumes to boot from",
   "Warning": "Warning",
