+++ src/views/virtualmachines/details/tabs/diagnostic/hooks/useDianosticData.ts
@@ -1,27 +1,24 @@
 import { useMemo } from 'react';
 
 import { V1VirtualMachine } from '@kubevirt-ui/kubevirt-api/kubevirt';
-import { columnSorting } from '@virtualmachines/list/hooks/utils/utils';
-import { PaginationState } from '@virtualmachines/utils';
 
-import { DiagnosticSort } from '../utils/types';
+import { VirtualizationStatusCondition, VirtualizationVolumeSnapshotStatus } from '../utils/types';
 import { conditionsTransformer, volumeSnapshotStatusesTransformer } from '../utils/utils';
 
-const useDiagnosticData = (
-  vm: V1VirtualMachine,
-  sort: DiagnosticSort,
-  pagination: PaginationState,
-) => {
+type UseDiagnosticData = (vm: V1VirtualMachine) => {
+  conditions: VirtualizationStatusCondition[];
+  volumeSnapshotStatuses: VirtualizationVolumeSnapshotStatus[];
+};
+
+const useDiagnosticData: UseDiagnosticData = (vm) => {
   const data = useMemo(() => {
     const volumeSnapshotStatuses = volumeSnapshotStatusesTransformer(
       vm?.status?.volumeSnapshotStatuses,
     );
     const conditions = conditionsTransformer(vm?.status?.conditions);
 
-    const combinedData = [...conditions, ...volumeSnapshotStatuses];
-
-    return columnSorting(combinedData, sort?.direction, pagination, sort?.column);
-  }, [vm, sort, pagination]);
+    return { conditions, volumeSnapshotStatuses };
+  }, [vm]);
 
   return data;
 };
